notificationprojet/pil1_2425_g13_backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
mysqlclient>=2.0,<3.0
django-cors-headers>=3.7,<4.0
djangorestframework-simplejwt>=5.0,<6.0